pyproject.toml
[tool.poetry]
name = "kpt-catalog-management"
version = "0.1.0"
description = ""
authors = ["ayushgarg0694 <ayushgarg0694@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.37"
alembic = "^1.8.0"
psycopg2 = "^2.9.3"
asyncpg = "^0.25.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.114.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
